ffs this was half-working for a bit and then i fucked it up and now i cant get it working at all again

+++ cv_outs.h
@@ -22,7 +22,7 @@ float clock_multiplier[NUM_CLOCKS] = {
 #include "sequencer.h"
 #endif
 
-int clock_delay[NUM_CLOCKS] = {
+byte clock_delay[NUM_CLOCKS] = {
   0, 0, 0, 0
 };
